pyproject.toml
[project]
name = "35-expense-tracking-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gspread>=6.2.1",
    "mysql-connector-python>=9.3.0",
    "oauth2client>=4.1.3",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "supabase>=2.15.3",
]
